requirements.txt
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg2-binary>=2.9,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
python-multipart>=0.0.9,<1.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
ocrmypdf>=16.0,<17.0
pdfminer.six>=20231228,<20260000
openai>=1.0,<2.0
anthropic>=0.39,<1.0
pymupdf>=1.24,<2.0
Pillow>=10.0,<11.0
pytesseract>=0.3,<1.0
bcrypt>=4.0,<5.0
python-jose[cryptography]>=3.3,<4.0
email-validator>=2.0,<3.0
